Optimize the uninstall check and update tests
