--- tests/phpunit/testdata/plugins/test-plugin-uninstall-constant-without-errors/load.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Plugin Name: Test Plugin Uninstall Constant without Errors for Plugin Check
- * Plugin URI: https://github.com/WordPress/plugin-check
- * Description: Some plugin description.
- * Requires at least: 6.0
- * Requires PHP: 5.6
- * Version: 1.0.0
- * Author: WordPress Performance Team
- * Author URI: https://make.wordpress.org/performance/
- * License: GPLv2 or later
- * License URI: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
- * Text Domain: test-plugin-uninstall-constant-without-errors
- * Update URI: https://wordpress.org/plugins/example-plugin/
- *
- * @package test-plugin-uninstall-constant-without-errors
- */
