--- tests/phpunit/testdata/plugins/test-plugin-uninstall-constant-without-errors/uninstall.php
@@ -1,4 +1,0 @@
-<?php
-if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
-	die;
-}
